ref(bot): separate branch protections test
